+++ src/platform/windows/sound/sound.cpp
@@ -4,23 +4,23 @@
 namespace Sound {
 std::deque<uint16_t> buffer;
 std::mutex audioMutex;
-};
+};  // namespace Sound
 
 namespace {
 SDL_AudioDeviceID dev = 0;
 
 void audioCallback(void* userdata, Uint8* raw_stream, int len) {
-    for (int i = 0; i<len; i++) raw_stream[i] = 0;
+    for (int i = 0; i < len; i++) raw_stream[i] = 0;
 
     std::unique_lock<std::mutex> lock(Sound::audioMutex);
-    int bufSize = Sound::buffer.size();
-    for (int i = 0; i<len; i+=2) {
-        if (i/2 >= bufSize) break;
+    size_t bufSize = Sound::buffer.size();
+    for (size_t i = 0; i < (size_t)len; i += 2) {
+        if (i / 2 >= bufSize) break;
 
         int16_t sample = Sound::buffer.front();
         Sound::buffer.pop_front();
         raw_stream[i] = sample & 0xff;
-        raw_stream[i+1] = (sample>>8) & 0xff;
+        raw_stream[i + 1] = (sample >> 8) & 0xff;
     }
 }
 }  // namespace
@@ -53,6 +53,4 @@ void Sound::stop() { SDL_PauseAudioDevice(dev, true); }
 
 void Sound::close() { SDL_CloseAudioDevice(dev); }
 
-void Sound::clearBuffer() {
-    buffer.clear();
-}
+void Sound::clearBuffer() { buffer.clear(); }
