+++ src/utils/psf.cpp
@@ -28,7 +28,7 @@ bool loadExe(System* sys, const std::vector<uint8_t>& _exe, PsfType type) {
         return false;
     }
 
-    for (size_t i = 0; i < exe.t_size; i++) {
+    for (uint32_t i = 0; i < exe.t_size; i++) {
         sys->writeMemory8(exe.t_addr + i, _exe[0x800 + i]);
     }
